rename pkg model to types, upgrade go version
