+++ README.md
@@ -1,1 +1,3 @@
 # esphome_api
+A Go library to manage [ESPHome](https://esphome.io/) devices.<br>
+To know more about usage look into [examples](/examples/) directory
