+++ examples/commom.go
