+++ examples/tail_logs/main.go
@@ -6,17 +6,17 @@ import (
 	"log"
 	"time"
 
-	example "github.com/mycontroller-org/esphome_api/example"
-	"github.com/mycontroller-org/esphome_api/pkg/model"
+	examples "github.com/mycontroller-org/esphome_api/examples"
+	types "github.com/mycontroller-org/esphome_api/pkg/types"
 	"google.golang.org/protobuf/proto"
 )
 
 func main() {
-	var logLevel = flag.Int("log_level", int(model.LogLevelVeryVerbose), "log level (0-6)")
+	var logLevel = flag.Int("log_level", int(types.LogLevelVeryVerbose), "log level (0-6)")
 	var monitoringDuration = flag.Duration("monitoring_duration", 30*time.Second, "monitoring duration")
 	flag.Parse()
 
-	client, err := example.GetClient(handlerFuncImpl)
+	client, err := examples.GetClient(handlerFuncImpl)
 	if err != nil {
 		log.Fatalln(err)
 	}
@@ -24,7 +24,7 @@ func main() {
 
 	fmt.Println("Listening logs, will terminate in", *monitoringDuration)
 
-	err = client.SubscribeLogs(model.LogLevel(*logLevel))
+	err = client.SubscribeLogs(types.LogLevel(*logLevel))
 	if err != nil {
 		log.Fatalln(err)
 	}
@@ -33,7 +33,7 @@ func main() {
 }
 
 func handlerFuncImpl(msg proto.Message) {
-	log, err := model.GetLogEntry(msg)
+	log, err := types.GetLogEntry(msg)
 	if err != nil {
 		fmt.Printf("MSG: %+v\n", msg)
 	}
